Rename async.py to async_.py for Python 3.7 compatibility
async became a reserved keyword in Python 3.7.
